fix: change loop index type from int64_t to size_t for consistency and cast result index to int64_t

+++ cxx/include/pyinterp/pybind/descriptive_statistics.hpp
@@ -350,8 +350,8 @@ auto DescriptiveStatistics<T>::calculate_statistics(Args... args) const
   const auto n = size();
   Eigen::VectorX<ResultType> result(n);
 
-  for (int64_t ix = 0; ix < n; ++ix) {
-    result[ix] = (accumulators_[ix].*MemberFunc)(args...);
+  for (size_t ix = 0; ix < n; ++ix) {
+    result[static_cast<int64_t>(ix)] = (accumulators_[ix].*MemberFunc)(args...);
   }
   return result;
 }
